+++ changes.md
@@ -1,8 +1,16 @@
 **Api changes**
 
 <details>
-<summary>Changed Property(s)</summary>
+<summary>Added Type(s)</summary>
 
-- :warning: changed property `discount` of type `DiscountedTotalPricePortion` from type `CartDiscountReference` to `Reference`
+- added type `ShoppingListSetBusinessUnitAction`
+</details>
+
+
+<details>
+<summary>Added Property(s)</summary>
+
+- added property `businessUnit` to type `ShoppingList`
+- added property `businessUnit` to type `ShoppingListDraft`
 </details>